backend/requirements.txt
# ===========================
# Zero-Touch Onboarding Backend
# ===========================

# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# AI/ML
openai==1.12.0
anthropic==0.18.1
chromadb==0.4.22

# Document Processing
pandas==2.2.0
pymupdf==1.23.22
python-docx==1.1.0

# Google APIs
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
